+++ src/lib/index.js
@@ -15,4 +15,5 @@ export * from './share.js';
 export * from './quality.js';
 export * from './impact.js';
 export * from './heatmap.js';
+export * from './graph-layout.js';
 export * from './remediation.js';